renamed toggleEditingId and handleToggleEditingId. Rewrote function toggleEditingId
