+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Layout } from "antd";
 import "./App.scss";
 import { Content } from "antd/es/layout/layout";
-import { RoutesList } from "./components/Routing/RoutesList";
+import { RoutesList } from "./shared/RoutesList";
 import { SideMenu } from "./components/SideMenu/SideMenu";
 
 const App = () => {
