+++ src/shared/RoutesList.tsx
@@ -1,7 +1,7 @@
 import { Route, Routes } from "react-router-dom";
-import { ProfilePage } from "../../pages/ProfilePage/ProfilePage";
-import { TodosPage } from "../../pages/TodosPage/TodosPage";
-import { NotFoundPage } from "../../pages/NotFoundPage/NotFoundPage";
+import { ProfilePage } from "../pages/ProfilePage/ProfilePage";
+import { TodosPage } from "../pages/TodosPage/TodosPage";
+import { NotFoundPage } from "../pages/NotFoundPage/NotFoundPage";
 
 export const RoutesList = () => {
     return (
